I created a new folder for elements and  moverd the buttun folder to it and created a new folder to make a separated component for the logo
