+++ next-project/src/app/globals.css
@@ -86,6 +86,7 @@
   box-sizing: border-box;
   padding: 0;
   margin: 0;
+  font-family: 'Montserrat';
 }
 
 html,
