+++ next-project/src/elements/Button/Button.module.css
@@ -1,0 +1,3 @@
+.logout {
+  background-color: aqua;
+}
