+++ next-project/src/elements/Logo/Logo.module.css
@@ -1,0 +1,5 @@
+.logo {
+  font-weight: bold;
+  font-size: 22px;
+  color: yellow;
+}
